// packages/headless/src/modules/system/recaptcha/types.d.ts
// --------------------------------------------------------
// ENUMS

// --------------------------------------------------------
// Contexts

export interface RecaptchaContext {
  grecaptcha?: any;
  token?: string;
  created?: Date;
  // ---
  error?: RequestError;
}

// --------------------------------------------------------
// Events

export interface RecaptchaEvent {
  type: string;
  data: any;
  error?: RequestError;
}